pkg/config: Split package file
