+++ pkg/config/config.go
@@ -1,7 +1,6 @@
 package config
 
 import (
-	"errors"
 	"strings"
 
 	"github.com/Xuanwo/storage/pkg/credential"
@@ -9,28 +8,6 @@ import (
 	"github.com/Xuanwo/storage/types"
 	"github.com/Xuanwo/storage/types/pairs"
 )
-
-var (
-	// ErrInvalidConfig will be returned when config is invalid.
-	ErrInvalidConfig = errors.New("invalid config")
-)
-
-// Error is the error returned by config
-type Error struct {
-	Op     string
-	Config string
-
-	Err error
-}
-
-func (e *Error) Error() string {
-	return e.Op + " " + e.Config + ": " + e.Err.Error()
-}
-
-// Unwrap implements xerrors.Wrapper
-func (e *Error) Unwrap() error {
-	return e.Err
-}
 
 // Parse will parse config string and return service type and namespace.
 func Parse(cfg string) (t string, opt []*types.Pair, err error) {
